refactor: reorganizar estructura de paquetes del backend siguiendo estándares de Spring Boot
✨ Reorganización completa de la arquitectura del backend:

📁 **Estructura de paquetes creada:**
- **config/** - Configuraciones (Security, Swagger, Web, Hibernate)
- **controller/** - Controladores REST (Product, Auth)
- **service/** - Lógica de negocio (Product, User)
- **model/** - Entidades JPA (Product, User)
- **repository/** - Acceso a datos (JPA y Hibernate nativo)
- **security/** - Seguridad JWT (Filter, TokenUtil)
- **exception/** - Excepciones personalizadas
- **util/** - Utilidades y DataLoader

🔧 **Cambios realizados:**
- ✅ Movimiento de 17 archivos Java a sus paquetes correctos
- ✅ Actualización de imports en todos los archivos
- ✅ Corrección de conflictos de nombres (Spring Security User)
- ✅ Actualización de tests para usar nuevos paquetes
- ✅ Mantenimiento de BackendApplication en la raíz

🧪 **Verificaciones completadas:**
- ✅ Compilación exitosa
- ✅ Tests unitarios pasando
- ✅ Construcción del JAR exitosa
- ✅ Estructura estándar de Spring Boot aplicada

Esta reorganización mejora la mantenibilidad, legibilidad y seguimiento
de las mejores prácticas de Spring Boot para proyectos empresariales.
