+++ backend/src/main/java/com/example/backend/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.example.backend;
+package com.example.backend.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
