+++ backend/src/main/java/com/example/backend/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.example.backend;
+package com.example.backend.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
