+++ backend/src/main/java/com/example/backend/controller/AuthController.java
@@ -1,4 +1,7 @@
-package com.example.backend;
+package com.example.backend.controller;
+
+import com.example.backend.security.JwtTokenUtil;
+import com.example.backend.service.UserService;
 
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
