+++ backend/src/main/java/com/example/backend/exception/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package com.example.backend;
+package com.example.backend.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
