+++ backend/src/main/java/com/example/backend/repository/UserRepository.java
@@ -1,4 +1,6 @@
-package com.example.backend;
+package com.example.backend.repository;
+
+import com.example.backend.model.User;
 
 import org.springframework.data.jpa.repository.JpaRepository;
