+++ backend/src/main/java/com/example/backend/security/JwtTokenUtil.java
@@ -1,4 +1,4 @@
-package com.example.backend;
+package com.example.backend.security;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
