+++ backend/src/test/java/com/example/backend/ProductRepositoryTest.java
@@ -1,5 +1,8 @@
 package com.example.backend;
 
+import com.example.backend.model.Product;
+import com.example.backend.repository.ProductRepository;
+
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
